2019-03-13 11:40
Finished Chapter 8 Testing; the React componont tests do not
work, probably because babel is missing a required dependency,
babel core 7 or something.  Need to debug further, but at
this point all the code seems to be correct and scripts work
correctly to update the package.json file and run the tests.