Version 0.0.6 2019-03-28 19:17
Removed gulp commands from evaluation, as they will not work
until after the files have been tangled; did not find a way to
tangle a single file from within code, but I suspect it is
possible because it works interactively.

The code now installs itself and starts the Snapkite-Engine
server with just the 'export' commnand.  The files tangle with
the 'tangle' command, 'C-c C-v C-t'.  Then, move into the
snapterest directory, and run the 'gulp' command to build the
snapterest file.  Then open 'index.html' in a browser, and the
React app works (even though the Twitter API is broken).