Version 0.0.7 2019-04-07 08:18
When tangling the files, Org first saves the source file,
reactjs-essentials.org, which has changed, at least in the macros
being expanded, and also apparently code blocks that have been
expanded and results written.  For some reason, these changes are
apparently reverted at the conclusion of the process in the
absence of the tangling.  If the source file is made read-only,
however, then the process works to completion, with the exception
that Emacs knows the source file has been changed on disk.

I am saving this change and then proceeding to find a work-around
for changed file on disk problem.